Changed code to avoid warnings from visual studio.
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%403423

+++ dlib/test/empirical_kernel_map.cpp
@@ -64,7 +64,7 @@ namespace
                 double err;
 
                 // the samples in the basis should have zero projection error
-                for (int i = 0; i < samples.size(); ++i)
+                for (unsigned long i = 0; i < samples.size(); ++i)
                 {
                     ekm.project(samples[i], err);
                     DLIB_TEST_MSG(abs(err) < 1e-13, abs(err));
